Move location schemas to a less awful location
Still not super happy with the place, but better than importing it from
the sessions module...
