+++ indico/modules/events/contributions/schemas.py
@@ -22,9 +22,10 @@ from indico.modules.events.contributions.models.references import ContributionRe
 from indico.modules.events.contributions.models.types import ContributionType
 from indico.modules.events.person_link_schemas import ContributionPersonLinkSchema as _ContributionPersonLinkSchema
 from indico.modules.events.sessions.models.blocks import SessionBlock
-from indico.modules.events.sessions.schemas import BasicSessionBlockSchema, BasicSessionSchema, LocationDataSchema
+from indico.modules.events.sessions.schemas import BasicSessionBlockSchema, BasicSessionSchema
 from indico.modules.events.tracks.schemas import TrackSchema
 from indico.modules.users.schemas import AffiliationSchema
+from indico.util.locations import LocationDataSchema
 from indico.util.marshmallow import EventTimezoneDateTimeField, SortedList
